tls: add TLSSocket.getEphemeralKeyInfo()
Returns an object representing a type, name and size of an ephemeral
key exchange(PFS) in a client connection.

This api is only supported not on a server connection but on a
client. When it is called on a server connection or its key exchange
is not ephemeral, an empty object is returned.