deps,openssl: add new make rule for asm of *.S
This is a floating patch against openssl-1.1.0 in order to have
asm files of *.S which are before pre-processed.
Node.js needs them in its repository because pre-processing is to be
made in the build of Node binaries on each architecture platforms.

This patch is to be submitted to the upstream.