build: terminate configure when old assembler
This can force users to know the requirements of a new assembler in
order to make the most use of crypto hardware features of CPU such as
AVX2. The requirements of assembler versions follow the forthcoming
OpenSSL-1.1.1 to support AVX512 and it avoids not to be semver-major
updates in the future.