deps: fix for MacOS and Win for OpenSSL-1.1.1
Because llvm on MacOS does not support AVX-512, asm files need to be
limited to AVX-2 support even when they are generated on Linux.
fake_gcc.pl returns the fake llvm banner version for MacOS as if the
assembler supports upto AVX-2.

For Windows, makefiles for nmake were updated in OpenSSL-1.1.1 and
they are rewritten into GNU makefile format by hand.